refactor actor msg

+++ src/qtype.h
@@ -5,8 +5,8 @@
 #ifndef __QTYPE_H__
 #define __QTYPE_H__
 
-typedef char*               qptr_t;
-typedef unsigned long long  qid_t;
+typedef char*         qptr_t;
+typedef unsigned long qid_t;
 
 #define QINVALID_ID 0
